SOFT30121-Advanced-Analysis-Design/src/test/java/com/example/firstsolutions/Controller/TrackControllerTest.java
package com.example.firstsolutions.Controller;

import org.junit.jupiter.api.Test;
import static org.junit.jupiter.api.Assertions.*;

class TrackControllerTest {


    @Test
    void testTrackItem() {
        // Assuming TrackItem has a constructor that initializes its properties
        TrackController.TrackItem trackItem = new TrackController.TrackItem("1", "Role", "Username", "ItemID", "Name", "5", "Message");

        // Now you can use its getters
        assertEquals("1", trackItem.getTrackID());
        assertEquals("Role", trackItem.getRole());
        assertEquals("Username", trackItem.getUsername());
        assertEquals("ItemID", trackItem.getItemID());
        assertEquals("Name", trackItem.getName());
        assertEquals("5", trackItem.getQuantity());
        assertEquals("Message", trackItem.getMessage());
    }

    @Test
    void testTrackFridgeActivity() {
        TrackController trackController = new TrackController();
        // Set up necessary dependencies if needed

        // Example test data
        String role = "Chef";
        String username = "john_doe";
        String itemID = "123";
        String name = "Tomato";
        String quantity = "5";
        String message = "Delivered";

        // Test the method
        trackController.trackFridgeActivity(role, username, itemID, name, quantity, message);

        // Add assertions for the expected behavior based on the method's logic
    }
}
